requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
openai>=1.40.0
tiktoken>=0.7.0
python-dotenv>=1.0.0